weather-app-backend/requirements.txt
django ~= 4.2
django-cors-headers ~= 4.3.1
djangorestframework ~= 3.14.0
apscheduler ~= 3.10.4